Updated and moved profile required decorator to student_profile app, and added no_profile_required decorator
